Finished porting most of the local tests (and split the tests back into multiple files)
