+++ cloudify_cli/tests/commands/tests/test_version.py
@@ -1,0 +1,7 @@
+from cloudify_cli.tests.commands.test_base import CliCommandTest
+
+
+class VersionTest(CliCommandTest):
+
+    def test_version(self):
+        self.invoke('cfy --version')
